SW-7569 Move EXIF logic to extension functions
The functions for extracting captured dates and geolocations aren't specific to
activity media files, so turn them into public extension functions.
